+++ packages/@expo/cli/src/utils/remote-build-cache-providers/helpers.ts
@@ -1,3 +1,5 @@
+import fs from 'fs';
+
 import { type Options as AndroidRunOptions } from '../../run/android/resolveOptions';
 import { type Options as IosRunOptions } from '../../run/ios/XcodeBuild.types';
 import { hasDirectDevClientDependency } from '../../start/detectDevClient';
@@ -21,4 +23,33 @@ export function isDevClientBuild({
   }
 
   return true;
+}
+
+export function moduleNameIsDirectFileReference(name: string): boolean {
+  // Check if path is a file. Matches lines starting with: . / ~/
+  if (name.match(/^(\.|~\/|\/)/g)) {
+    return true;
+  }
+
+  const slashCount = name.split('/')?.length;
+  // Orgs (like @expo/config ) should have more than one slash to be a direct file.
+  if (name.startsWith('@')) {
+    return slashCount > 2;
+  }
+
+  // Regular packages should be considered direct reference if they have more than one slash.
+  return slashCount > 1;
+}
+
+export function moduleNameIsPackageReference(name: string): boolean {
+  const slashCount = name.split('/')?.length;
+  return name.startsWith('@') ? slashCount === 2 : slashCount === 1;
+}
+
+export function fileExists(file: string): boolean {
+  try {
+    return fs.statSync(file).isFile();
+  } catch {
+    return false;
+  }
 }
